requirements.txt
pandas
requests
psycopg2-binary
google-cloud-bigquery
python-dotenv